requirements.txt
torch==1.9.0
torchvision==0.10.0
opencv-python==4.7.0.72
tqdm==4.51.0
einops==0.6.1
numpy==1.20.2
timm==0.4.12
Pillow==9.5.0
scipy==1.7.3
decord >= 0.4.1
mmcv==2.0.0 
mmaction2==1.0.0